+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ### 1.2.6
 - Fixed an issue with parsing PGNs containing clock annotations (thanks @deathcoder).
+- Variant: Antichess.
 - `Variant.forcedCapture`: defines forced capture rule behaviour.
 - `GameEndConditions` changes: `stalemate` and `elimination` are now of type `EndType` instead of `bool`, allowing greater flexibility (win, loss or draw).
-- Antichess variant.
 
 
 ### 1.2.5