+++ CHANGELOG.md
@@ -1,9 +1,11 @@
+### 1.2.7
+- `BishopState.capturedPieces()`: used to get the captured pieces in a given state, by comparing to the initial state.
+
 ### 1.2.6
 - Fixed an issue with parsing PGNs containing clock annotations (thanks @deathcoder).
 - Variant: Antichess.
 - `Variant.forcedCapture`: defines forced capture rule behaviour.
 - `GameEndConditions` changes: `stalemate` and `elimination` are now of type `EndType` instead of `bool`, allowing greater flexibility (win, loss or draw).
-
 
 ### 1.2.5
 - Improved efficiency when making moves from SAN strings or parsing PGNs: the check component of the SAN string can now be skipped.